requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4; platform_system == "Darwin"
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
argparse==1.4.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b7
azure-common==1.1.28
azure-core==1.32.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-cosmos==4.9.0
azure-identity==1.21.0
azure-monitor-opentelemetry==1.6.5
azure-monitor-opentelemetry-exporter==1.0.0b35
azure-search-documents==11.5.2
babel==2.17.0
beautifulsoup4==4.13.3
bleach[css]==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6; platform_system == "Windows" or sys_platform == "win32"
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
deprecated==1.2.18
distro==1.9.0
exceptiongroup==1.2.2; python_version < "3.11"
executing==2.2.0
fastjsonschema==2.21.1
fixedint==0.1.6
fqdn==1.5.1
gitdb==4.0.12
gitpython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==8.6.1
ipykernel==6.29.5
ipython==8.34.0
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.9.0
json5==0.10.0
jsonpointer==3.0.0
jsonschema-specifications==2024.10.1
jsonschema[format-nongpl]==4.23.0
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.3.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
markupsafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
msal==1.32.0
msal-extensions==1.3.1
msrest==0.7.1
narwhals==1.31.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.3
notebook-shim==0.2.4
numpy==2.2.4
oauthlib==3.2.2
openai==1.68.2
opentelemetry-api==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-dbapi==0.52b1
opentelemetry-instrumentation-django==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-flask==0.52b1
opentelemetry-instrumentation-psycopg2==0.52b1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-urllib==0.52b1
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-wsgi==0.52b1
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdfminer-six==20231228
pdfplumber==0.11.5
pexpect==4.9.0; sys_platform != "win32" and sys_platform != "emscripten"
pillow==11.1.0
platformdirs==4.3.7
prometheus-client==0.21.1
prompt-toolkit==3.0.50
protobuf==5.29.4
psutil==6.1.1
ptyprocess==0.7.0; sys_platform != "win32" and sys_platform != "emscripten" or os_name != "nt"
pure-eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydeck==0.9.1
pygments==2.19.1
pyjwt[crypto]==2.10.1
pymupdf==1.25.4
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2025.1
pywin32==310; sys_platform == "win32" and platform_python_implementation != "PyPy"
pywinpty==2.0.15; os_name == "nt"
pyyaml==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
send2trash==1.8.3
setuptools==77.0.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
streamlit==1.43.2
tenacity==9.0.0
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1; python_version < "3.11"
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
watchdog==6.0.0; platform_system != "Darwin"
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.2
zipp==3.21.0
